requirements.txt
absl-py==2.1.0
accelerate==0.21.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-cors==0.7.0
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==24.2.0
autogluon==1.1.1
autogluon.common==1.1.1
autogluon.core==1.1.1
autogluon.features==1.1.1
autogluon.multimodal==1.1.1
autogluon.tabular==1.1.1
autogluon.timeseries==1.1.1
beautifulsoup4==4.12.3
blis==0.7.11
boto3==1.35.41
botocore==1.35.41
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpathlib==0.19.0
cloudpickle==3.1.0
cmake==3.30.5
colorama==0.4.6
coloredlogs==15.0.1
colorful==0.5.6
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
datasets==2.14.4
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.9
evaluate==0.4.3
fastai==2.7.17
fastcore==1.7.17
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
ftfy==6.3.0
future==1.0.0
gdown==5.2.0
gluonts==0.15.1
google-api-core==2.21.0
google-auth==2.35.0
googleapis-common-protos==1.65.0
grpcio==1.67.0
huggingface-hub==0.25.2
humanfriendly==10.0
hyperopt==0.2.7
idna==3.10
imageio==2.36.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.21.1
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
lazy_loader==0.4
lightgbm==4.3.0
lightning==2.3.3
lightning-utilities==0.11.8
llvmlite==0.43.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
mlforecast==0.10.0
model-index==0.1.11
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.15
murmurhash==1.0.10
networkx==3.4.1
nlpaug==1.1.11
nltk==3.9.1
nptyping==2.4.1
numba==0.60.0
numpy==1.26.4
nvidia-ml-py3==7.352.0
omegaconf==2.2.3
onnx==1.17.0
onnxruntime==1.16.3
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.11
optimum==1.17.1
ordered-set==4.1.0
orjson==3.10.7
packaging==24.1
pandas==2.2.3
patsy==0.5.6
pdf2image==1.17.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
preshed==3.0.9
prometheus_client==0.21.0
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
psutil==5.9.8
py-spy==0.3.14
py4j==0.10.9.7
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.3
pytorch-metric-learning==2.3.0
pytz==2024.2
PyWavelets==1.7.0
PyYAML==6.0.2
ray==2.10.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.3
safetensors==0.4.5
scikit-image==0.20.0
scikit-learn==1.4.0
scipy==1.12.0
sentencepiece==0.2.0
seqeval==1.2.2
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
soupsieve==2.6
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
statsforecast==1.4.0
statsmodels==0.14.4
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
text-unidecode==1.3
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==0.9.16
tokenizers==0.19.1
toolz==0.12.1
torch==2.2.2
torchmetrics==1.2.1
torchvision==0.17.2
tqdm==4.66.5
transformers==4.40.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
utilsforecast==0.0.10
uv==0.4.22
virtualenv==20.26.6
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.0.4
window_ops==0.0.15
wrapt==1.16.0
xgboost==2.0.3
xxhash==3.5.0
yarl==1.15.3